+++ speculative_decoding/helpers.py
@@ -1,18 +1,25 @@
 import collections
 import contextlib
+import os
 import socket
+import subprocess
 import time
+from pathlib import Path
 from typing import Sequence, Union
 
 import numpy as np
 import pandas as pd
 import pydantic
 import tritonclient.grpc as triton_grpc
+import tritonclient.http as triton_http
 import tritonclient.utils as triton_utils
+
+# Timing.
+########################################################################################
 
 
 class _Timing(object):
-    """Accumulator that collects statistics (esp. execution time) over repeated updates and offers pretty printing."""
+    """Oject to collect code section timing statistics."""
 
     def __init__(self):
         self._values = collections.deque(maxlen=50)
@@ -37,15 +44,6 @@ class _Timing(object):
         self.update(t)
         self._t0 = 0
 
-    def __str__(self) -> str:
-        vals = np.array(self._values)
-        avg = self._sum / self._count
-        its = 1 / avg
-        return (
-            f"Σ={self._sum:.3g} ({its:.3g}/s)\tμ={avg:.3g}\tmin={vals.min():.3g}\t"
-            f"max={vals.max():.3g}\tσ/μ={vals.var() / avg:.3g}\tN={self._count}"
-        )
-
     def to_dict(self) -> dict[str, float | int | str]:
         vals = np.array(self._values)
         avg = self._sum / self._count
@@ -61,12 +59,13 @@ class _Timing(object):
 
 
 class _TimingManager(object):
-    """Container to manage a set of global metrics for execution times."""
+    """Object to manage collection of `_Timing`s."""
 
     def __init__(self):
         self._timings = collections.defaultdict(_Timing)
         self._order = []
         self._total = None
+        self.enabled = False
 
     def start_time(self, name: str):
         if name not in self._timings:
@@ -75,19 +74,6 @@ class _TimingManager(object):
 
     def record_time(self, name: str):
         self._timings[name].record_time()
-
-    def __str__(self):
-        parts = ["Global Wall Time Totals:"]
-        for n in self._order:
-            m = self._timings[n]
-            if self._total:
-                t = self._timings[self._total].sum
-                fraction = f"    {m.sum / t * 100:2.0f}% "
-            else:
-                fraction = ""
-
-            parts.append(f"{n:<20}: {fraction}{m}")
-        return "\n".join(parts)
 
     def to_pd(self) -> pd.DataFrame:
         rows = []
@@ -100,10 +86,9 @@ class _TimingManager(object):
         return pd.DataFrame(data=rows, columns=columns).set_index(section_key)
 
     def reset(self):
-        self.__init__()
-
-    def set_total(self, name: str):
-        self._total = name
+        self._timings.clear()
+        self._order.clear()
+        self._total()
 
 
 _global_metrics = _TimingManager()
@@ -111,17 +96,29 @@ _global_metrics = _TimingManager()
 
 @contextlib.contextmanager
 def timeit(name: str):
-    try:
-        _global_metrics.start_time(name)
+    if _global_metrics.enabled:
+        try:
+            _global_metrics.start_time(name)
+            yield
+        finally:
+            _global_metrics.record_time(name)
+    else:
         yield
-    finally:
-        _global_metrics.record_time(name)
+
+
+def enable_timing() -> None:
+    _global_metrics.enabled = True
+
+
+def disable_timing() -> None:
+    _global_metrics.enabled = False
 
 
 def show_timings():
     print(_global_metrics.to_pd())
 
 
+# Request API.
 ########################################################################################
 
 
@@ -149,6 +146,10 @@ class GenerationRequest(pydantic.BaseModel):
     bad_word_list: Sequence[str] | None = None
     stop_words_list: Sequence[str] | None = None
     sampling_config: SamplingConfig = SamplingConfig()
+
+
+# Triton inference client helpers.
+########################################################################################
 
 
 def _fill_inputs(
@@ -229,3 +230,95 @@ def extract_trtllm_outputs(result: triton_grpc.InferResult) -> np.ndarray[np.int
     )
     assert sequence_len == len(output_ids), f"{sequence_len} vs. {len(output_ids)}"
     return output_ids
+
+
+# Triton inference server helpers.
+########################################################################################
+
+GRPC_SERVICE_PORT = 8001
+HTTP_SERVICE_PORT = 8003
+
+
+def is_triton_server_alive() -> bool:
+    def port_is_available(port):
+        available = False
+        with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
+            try:
+                sock.bind(("0.0.0.0", port))
+                available = True
+            except OSError:
+                pass
+        return available
+
+    return not port_is_available(GRPC_SERVICE_PORT) or not port_is_available(
+        HTTP_SERVICE_PORT
+    )
+
+
+class TritonServer:
+    def __init__(self, model_repository_dir: Path, parallel_count=1):
+        self._model_repository_dir: Path = model_repository_dir
+        self._parallel_count = parallel_count
+
+    def load_server_and_model(self, env: dict) -> None:
+        """Loads the Triton server and the model."""
+        if not is_triton_server_alive():
+            self._start_server(mpi=self._parallel_count, env=env)
+
+        http_client = triton_http.InferenceServerClient(
+            url=f"localhost:{HTTP_SERVICE_PORT}", verbose=False
+        )
+
+        is_server_up = False
+        while not is_server_up:
+            try:
+                is_server_up = http_client.is_server_live()
+            except ConnectionRefusedError:
+                time.sleep(2)
+                continue
+
+        http_client.is_server_ready
+
+        while not http_client.is_model_ready(model_name="target_model"):
+            time.sleep(2)
+            continue
+
+    def build_server_start_command(self, mpi: int = 1, env: dict = {}) -> list:
+        base_command = [
+            "tritonserver",
+            "--model-repository",
+            str(self._model_repository_dir),
+            "--grpc-port",
+            str(GRPC_SERVICE_PORT),
+            "--http-port",
+            str(HTTP_SERVICE_PORT),
+        ]
+
+        if mpi == 1:
+            return base_command
+
+        mpirun_command = ["mpirun", "--allow-run-as-root"]
+        # Generate mpi_commands with a unique shm-region-prefix-name for each MPI process
+        mpi_commands = []
+        for i in range(mpi):
+            mpi_command = [
+                "-n",
+                "1",
+                *base_command,
+                "--disable-auto-complete-config",
+                f"--backend-config=python,shm-region-prefix-name=prefix{str(i)}_",
+            ]
+            mpi_commands.append(" ".join(mpi_command))
+
+        # Join the individual mpi commands with ' : ' as required by mpirun syntax for multiple commands
+        combined_mpi_commands = " : ".join(mpi_commands)
+
+        return mpirun_command + [combined_mpi_commands]
+
+    def _start_server(self, mpi: int = 1, env: dict = {}) -> subprocess.Popen:
+        """Triton Inference Server has different startup commands depending on
+        whether it is running in a TP=1 or TP>1 configuration. This function
+        starts the server with the appropriate command."""
+
+        command = self.build_server_start_command(mpi, env)
+        return subprocess.Popen(command, env={**os.environ, **env})
