+++ speculative_decoding/triton_model_repo/draft_model/config.pbtxt
@@ -215,7 +215,7 @@ parameters: {
   key: "gpt_model_path"
   value: {
     # string_value:  "/root/workbench/gpt2_engine/"
-    string_value:  "/packages/triton_model_repo/draft_model/1/"
+    string_value:  "/triton_model_repo/draft_model/1/"
   }
 }
 parameters: {
