--- speculative_decoding/triton_model_repo/dummy_model/config.pbtxt
@@ -1,24 +1,0 @@
-name: "dummy_model"
-backend: "python"
-max_batch_size: 2048
-input [
-  {
-    name: "input"
-    data_type: TYPE_INT32
-    dims: [ -1 ]
-  }
-]
-output [
-  {
-    name: "output"
-    data_type: TYPE_INT32
-    dims: [ -1 ]
-  }
-]
-
-instance_group [
-    {
-        count: 1
-        kind: KIND_CPU
-    }
-]
